Create README.md
The idea of connecting 'things' or physical objects has gained traction in recent years.  Under the Internet of Things(IoT) concept, objects with sensors are connected to the internet so as to share data and information with other objects over the internet.  One of the many areas that IoT is being applied is in healthcare.  For instance, doctors can use IoT devices to know whether patients are adhering to the medication provided. It is on this premise that this project is based; having the heart rate of a person recorded and the data streamed to Google cloud. A doctor or a health care provider can be able to monitor the heart rate of the person and give personalized attention. This is important especially when the person lives alone.